#7 Creation of Following, Followers, MyBuisinesses, and Friends List
After creating list components for the My Businesses , user Following, and User Friends, i merged with explore page and main page, I sectioned out the list files. I made a new folder for lists and a styling folder in that folder for css.

These lists included fetching of all businesses a uyser owns, all friends of a specific user, all businesses a user follows, and all followers of a business. Each containing a search nar that queries through the information that is feched and listed.