#7 #1 Added live follow updating fucntionality to Business Profile
A new addition of sockets has been added to the business Profile so that followers count auto updates on a user following and unfollowing another user.

Also changed socket events in userprofile and businessprofile to connect to a socket namespace that I created inside of app.jsx for both when looking through user profile or business profile.