#7 #5 Added veiw friends calendar button
I used the useRef hook in react to create a view calendar page for when a friend of a user wants to look at their calendar. The user should be able to add the event to their calendar, but that functionality has not been added yet.

In the user profile a render function that displayed the view calendar depending on the person viewing the page being the owner or are friends with them that links to the userCalendarView page.

I also added the route to the user calendar view in app.jsx so that is able to be linked to in in userCalendar View