#7 #4 Made function to add a calendarItem to personal Calendar
In the Home.jsx i changed the way the calendarItems were being called, and changed it to axios as there was an error coming from the routing and it was hard to follow.

In event detail modal, I added context for getting user information

I also added ocnditionals for when a user "deletes" a calendar item as well as buttons showing for who has permission to edit or publish/unpublish an event.