#7  Conditionals added and Message pop ups added as well
If a user is not th eowner of an event that they have on their calendar, they should not be able to see the publish/unpublish button on the event detail modal.

In the user calendar View, I added a state that will show a message that lets the user know if the addition of a calendar item to their calendar was successful.